# Conflicts:
#	src/main/java/com/nsa/group6/web/FormController.java
#	src/main/resources/static/css/reflections.css
#	src/main/resources/templates/form-view.html
#	src/main/resources/templates/form.html
#	src/main/resources/templates/home.html
#	src/main/resources/templates/reflection-list.html

+++ src/main/java/com/nsa/group6/service/FormService.java
@@ -13,7 +13,7 @@ public interface FormService {
 
     public void updateForm(Form aForm);
 
-    public void deleteForm(Long id);
+    public void deleteForm(int id);
 
     List<Form> getAllFormsByUsername(User username);
